crypto: Move types::Signature to its own module
